local/pyproject.toml
[tool.poetry]
name = "local"
version = "0.1.0"
description = ""
authors = ["B4rtware <34386047+B4rtware@users.noreply.github.com>", "Re-Krass <38668040+Re-Krass@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyzmq = "^23.1.0"
pydantic = "^1.9.1"
tornado = "^6.1"
SQLAlchemy = "^1.4.39"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
sqlalchemy2-stubs = "^0.0.2-alpha.24"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
